+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Devel
 
-- Support Django 4.1.*
+- Support Django 4.1.*.
+- Fix broken links in group membership deatil pages.
 
 ## 0.11 (2023-02-24)