+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support Django 4.1.*.
 - Fix broken links in group membership detail pages.
 - Add last update date to GroupGroupMembership, GroupAccountMembership, and WorkspaceGroupSharing tables.
+- Load workspace data form media in workspace editing templates.
 
 ## 0.11 (2023-02-24)