+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix broken links in group membership detail pages.
 - Add last update date to GroupGroupMembership, GroupAccountMembership, and WorkspaceGroupSharing tables.
 - Load workspace data form media in workspace editing templates.
+- Bugfix: Accessible workspaces for an account only includes workspaces that are shared with groups they are in.
 
 ## 0.11 (2023-02-24)