+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix broken links in group membership detail pages.
 - Add last update date to GroupGroupMembership, GroupAccountMembership, and WorkspaceGroupSharing tables.
 - Load workspace data form media in workspace editing templates.
+- Add a view to show a graph-based visualization of ManagedGroups.
+- Show a graph-based visualization of group relationships in the ManagedGroupDetail view.
 
 ## 0.11 (2023-02-24)